+++ autotest.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+TEST_ROOT_PATH="./test" pub run test