+++ lib/annotation.dart
@@ -1,23 +1,33 @@
 library params.annotation;
 
-import 'dart:js';
-export 'dart:js';
+abstract class Accessor<V1,V2> {
+  V1 get(V2 value);
+  V2 set(V1 value);
+}
 
-abstract class Converter {
-
+abstract class JSONEncoder{
+  dynamic encode(value);
 }
 
 abstract class Serialize{
   dynamic get $container$;
   dynamic operator[](property)  => $container$[property];
   operator[]=(property, dynamic value) => $container$[property] = value;
-  Map<String, dynamic> asMap() => throw new UnimplementedError('Use a transformer');
-  JsObject asJsObject() => throw new UnimplementedError('Use a transformer');
+  Map<String, dynamic> toJson([Map<String, dynamic> _$$result_$$_]){
+    if (_$$result_$$_ == null) {
+      throw new UnimplementedError('Use a transformer');
+    }
+    return _$$result_$$_;
+  }
 }
 
 
 class ModelParameter {
-  final Type converter;
+  final Type accessor;
   final String name;
-  const ModelParameter({this.name, this.converter});
+  final String json;
+  final bool jsonSkipEmpty;
+  final Type jsonEncoder;
+  final bool isSubModel;
+  const ModelParameter({this.name, this.accessor, this.json, this.isSubModel:false, this.jsonSkipEmpty: true, this.jsonEncoder});
 }