+++ lib/src/accessor_info.dart
@@ -1,16 +1,16 @@
 part of params.transformer;
 
-class ConverterInfo{
+class AccessorInfo{
   static int count = 0;
   final String className;
   final int index = count++;
 
-  String get field => "_\$converter_$index\$_";
+  String get field => "_\$accessor_$index\$_";
 
   String generateStaticFinalField(){
     return "static final $className $field = new $className();";
   }
 
-  ConverterInfo(this.className);
+  AccessorInfo(this.className);
   toString()=>"$className[$field]";
 }