+++ lib/src/json_encoder_registry.dart
@@ -1,0 +1,8 @@
+part of params.transformer;
+
+class JSONEncoderRegistry{
+  final Map<String, JSONEncoderInfo> _items = {};
+  JSONEncoderInfo get(String converterName){
+    return _items.putIfAbsent(converterName,()=>new JSONEncoderInfo(converterName));
+  }
+}