+++ test/sample/test1.in.dart
@@ -1,0 +1,8 @@
+library params.test;
+
+import 'package:params/annotation.dart';
+
+class SimpleTest1 extends Object with Serialize {
+  @ModelParameter()
+  String field1;
+}